refactor: parallelization
