+++ octopus/modules/autogluon/core.py
@@ -26,8 +26,8 @@ from sklearn.base import BaseEstimator, ClassifierMixin, RegressorMixin
 
 from octopus.experiment import OctoExperiment
 from octopus.logger import LogGroup, get_logger
+from octopus.manager.ray_parallel import setup_ray_for_external_library
 from octopus.modules.autogluon.module import AutoGluon
-from octopus.modules.octo.ray_parallel import setup_ray_for_external_library
 from octopus.modules.utils import (
     get_fi_group_shap,
     get_fi_shap,
