+++ octopus/modules/octo/bag.py
@@ -18,12 +18,12 @@ from sklearn.base import BaseEstimator, ClassifierMixin, RegressorMixin
 
 # sklearn imports for compatibility
 from octopus.logger import LogGroup, get_logger
-from octopus.metrics.utils import get_performance_from_predictions
 
 # Adjust this import path as needed depending on your package layout
-from octopus.modules.octo.ray_parallel import (
+from octopus.manager.ray_parallel import (
     run_parallel_inner,
 )
+from octopus.metrics.utils import get_performance_from_predictions
 from octopus.modules.octo.training import Training
 
 logger = get_logger()
